[FEATURE] add forwards for edit_distance
